feat: tests/test_utils.py の主要ヘルパーを jpoke.testing として本体パッケージへ昇格
pip install jpoke だけで（clone せずに）start_battle/run_move 等のテスト
ヘルパーとCustomPlayerを使えるようにする。tests/test_utils.py は後方互換の
薄い再エクスポート層に変更。移植にあたり内部マネージャー直呼びだった箇所は
Battle.set_ailment() へのsource/overwrite引数追加、Battle.set_volatile() /
set_item() / activate_global_field() / activate_side_field() /
resolve_action_order() / calc_move_priority() / can_switch() / end_turn()
の新設で公開API経由に置き換えた。fix_damage/fix_random はテスト専用の
モンキーパッチである旨をdocstringに明記のうえ据え置き。
